requirements.txt
streamlit==1.44.1
ultralytics==8.3.89
opencv-python==4.11.0.86
numpy==2.1.1
pandas==2.2.3
pillow==11.1.0
matplotlib==3.10.1
jupyter==1.1.1
